Finish soup.html and fix minor visual issues
In addition to completing the soup page, I added a couple colons to each
page as thry were all missing them. This makes it look a lot cleaner
and I went ahead and fixed the missing list items for the unordered
list on the main page. Ready for CSS styling...